Arc Plotter & Code Rewrite
This is a pretty big change. The Arc Plotter at least cheats correctly even though it's not properly setup with a Zingl-Bresenham plot routine. It converts to some number of cubics. Zingl stuff was put into the `path.py` and that was all rewritten. Some parts broke, like moving things.