Centralized Elements
This code likely spawns some additional bugs. The idea behind this is to properly centralize the location for selection and manipulation of objects so when the bounds need to be updated they can be proven valid and do not need recalculation. The same is true for selection of objects etc.