Auto-create operations for any encountered colors & enable context-menu conversion
Currently there's no way to "Add" an operation, and if you have paths that don't match up color-wise with operations, they are just binned as raster. This makes it so you can define different paths as different colors, and each color gets its own operation.

I also implemented the "Convert Operation" context menu, for ease of conversion.

Test plan:
open an SVG with an orange line, a blue line, and a red line.
the blue and red lines should get binned into the preexisting operations, and a new orange operation should be created.

right-click the orange auto-created "engrave" operation, use "Convert Operation" to change it to a `Cut`. Double-click on it to see that in the edit window it also appears as a "Cut".