Added Native DPI scaling for images.
If an image has a given DPI it will now load correctly and use that internal scaling. This is not a native resolution of the machine and it will have to resize it. This means it will have to actualize it too. I added a setting to turn this off.