Experimental Code
Added in some fake Rect object direct, and a BlobOperation and a Blob object as elements to sort out what is and isn't reasonable. Also, added in a gcode reader bit of code, but it forces the code to be applied directly to the spooler. Hence the extra work on blobs and bloboperations.