Created plan command, started adapting.
The need for a plan command is to do all the work of the jobs dialog within the console. This should do the various phases. Copy, Validate, Blob, Optimize, and Spool. These are copying the operations, validating those operations, converting those operations to a cutcode blob, optimizing that blob, and spooling the blob.