Fix incorrect error types loading drivers
ch341.py expects drivers to raise ImportError when a failure occurs during initialization, and ConnectionRefusedError when the failure occurs during connection.

libusb and windll both are throwing ConnectionRefusedError during the initialization phase. A failure to load the driver during this phase will not be caught by the encompassing code and will result in a failure to fall back to any alternate driver present.