Improve classification code for fill-only
As you can see, this is a rewrite of the classification code which ends up a bit simpler.

The Dots code has been commented out as there is AFAIK no support for a Dots operation elsewhere.

I will upload the test file I used in the comments. Inkscape cannot produce Dots as Move + Close, so I created two lines with two coincident nodes and edited the XML to create the Move + Close.