Create elements.py
Execute Job was not working - but no error messages

I traced this to the use of bootstrapped node types when the copied laser operation was not in the tree, and therefore did not have self._root set.

It didn't help that Execute Job (which works as usual by executing console tasks) had some sort of global exception handler which threw away the error messages!!!!! So one step in the Execute Job was getting an Exception, but we never knew. This is **INCREDIBLY** poor style. We should probably:

1. Have a Console Error exception class, and any explicit exceptions should raise this and be reported to the console, whilst any other exceptions should create an error report as usual.
2. Handle errors for user issued console commands differently from internally issued console commands - dont submit user generated errors centrally, just report them locally, whilst reporting centrally all errors during internal use of console commands.