Colorise operations icons
Allows operation color to be applied to the icons, replacing black with the color. Grey scale is handled by transparency.

However, if the operation color is close to the tree background color (assumed to be white in normal mode and black in dark mode) this will not contrast well.  So we replace the transparency with a white or black background and removing the alpha channel.