Create assets for pull_requests / pushes
It would be great to build executables for pull-requests so that we can test PRs as executables rather than just running from source.

I am building here on the excellent work done by tiger in the existing ci.yml.

This is an interim script which has releases commented out but the intent will be to replace the existing ci.yml once it is fully working.