Do not send crash details if running from git
The idea here is to avoid developers mistakenly sending crash reports when developing.

But it also prevents sending crash reports for production use of unchanged code.

On balance I think this is the right thing to do (otherwise I would not be putting it forwards), but others may disagree and I will go with the consensus.