Working interim version of multi-image raster generation
Generates multiple images just fine even for layered raster emblems.

Not sure whether we need to set the path property on the image objects - seem to remember this is used for cut inside first.

Still need to optimise for rasters which are extremely close to one another (within e.g. overscan) and optimise for time.