Improve menu key
Partial improvement if scene is active.
1. Menu key is now not a modifier but a key
2. If scene is active pane, then menu key is equivalent to right click on selected object - a menu key-down/up event is translated into a call to the on_right_down/up method.

Same does not work for tree, because key events are handled in the tree and mouse events in the window (which doesn't seem right to me) - anyway a key event cannot call a mouse method at a higher level because there is no object reference available.