Replace TextCtrl with RichTextCtrl
This commit simply provides a like-for-like replacement of the output text control with a RichText equivalent. Aside from a hanging indent for long lines, this simply duplicates the functionality of the existing TextCtrl.

This is a prerequisite for delivering colour coded Console output to both GUI-based console, and hopefully similar coloured output to the GUI suppressed command console loop and channel print output to console.