Make timeout for EPP-write longer in libusb.py
Moshiboard has a 32kB buffer on the board. The CH341 on the moshiboard makes the PC waits when the buffer is full until the buffer become available. This waiting time might be long enough to make device.write() in the line 440 of libusb.py give timeout exception; 
            device.write(endpoint=BULK_WRITE_ENDPOINT, data=data, timeout=self.timeout)
The waiting time can be easily long. For example, it is three seconds if the laser travels 30mm at 10mm/s. 

So this commit changes the timeout for EPP write to be 60 seconds.